// src/app/models/citizen-filtering-enum.ts
export enum CitizenFilter {
  taxesPayed = "✓ Tasse pagate",
  taxesUnpayed = "✕ Tasse non pagate",
  lowPerformance = "⚠ Cittadini da richiamare (performance bassa)",
  noFilter = "Nessun filtro"
}

export enum CitizenSorting {
  performanceDescending = "↘ Performance: alta-bassa",
  performanceAscending = "↗ Performance: bassa-alta",
  alphabetical = "↓ Alfabetico"
}